Split Anthropic model registry modules
